+++ src/index.ts
@@ -2,5 +2,4 @@ export * from './types';
 export * from './defaults';
 export * from './static';
 export * from './themed';
-export * from './props';
 export * from './components';