+++ CHANGELOG.md
@@ -9,9 +9,9 @@
 #### Additions
 
 - Increased the supported version range for `styled-components` to support `v4`
-- added aliases for the longer form of the mixin names e.g. for there `mt` you can now also use `marginTop`
-- added `createMarginTop`, `createPaddingTop` and similar methods for creating mixins for breakpoints not defined by the theme
+- added aliases for the longer form of the mixin names e.g. for there `mt` you can also use `marginTop` now
 - added `marginProps` and `paddingProps` methods for creating `Box`-like components
+- added `createMarginTop`, `createPaddingTop` and similar factories for creating mixins when using the `createMap` functions
 - Added a license file
 
 ## Deprecations